Refactor to branching approach - separated common code to `common.ts` - created `main-browser.ts` and connected via `package.json` "browser" field. - cleaned up never-read `resolvedLanguage` variable (only set values to it)
