+++ src/main.ts
@@ -2,17 +2,32 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
-'use strict';
-
 import * as path from 'path';
 import * as fs from 'fs';
+import {
+	format,
+	localize,
+	isDefined,
+	setPseudo,
+	isPseudo,
+	MessageFormat,
+	BundleFormat,
+	Options,
+	LocalizeInfo,
+	TranslationConfig,
+	LanguageBundle,
+	LocalizeFunc,
+	NlsBundle,
+	MetaDataFile,
+	MetadataHeader,
+	I18nBundle,
+	SingleFileJsonFormat,
+	LoadFunc
+} from './common';
+
+export { MessageFormat, BundleFormat, Options, LocalizeInfo, LocalizeFunc, LoadFunc, KeyInfo } from './common';
 
 const toString = Object.prototype.toString;
-const readFileSync = fs.readFileSync;
-
-function isDefined(value: any): boolean {
-	return typeof value !== 'undefined';
-}
 
 function isNumber(value: any): value is number {
 	return toString.call(value) === '[object Number]';
@@ -27,89 +42,7 @@ function isBoolean(value: any): value is boolean {
 }
 
 function readJsonFileSync<T = any>(filename: string): T {
-	return JSON.parse(readFileSync(filename, 'utf8')) as T;
-}
-
-export enum MessageFormat {
-	file = 'file',
-	bundle = 'bundle',
-	both = 'both'
-}
-
-export enum BundleFormat {
-	// the nls.bundle format
-	standalone = 'standalone',
-	languagePack = 'languagePack'
-}
-
-export interface Options {
-	locale?: string;
-	cacheLanguageResolution?: boolean;
-	messageFormat?: MessageFormat;
-	bundleFormat?: BundleFormat;
-}
-
-export interface LocalizeInfo {
-	key: string;
-	comment: string[];
-}
-
-namespace LocalizeInfo {
-	export function is(value: any): value is LocalizeInfo {
-		let candidate = value as LocalizeInfo;
-		return candidate && isDefined(candidate.key) && isDefined(candidate.comment);
-	}
-}
-
-export interface LocalizeFunc {
-	(info: LocalizeInfo, message: string, ...args: (string | number | boolean | undefined | null)[]): string;
-	(key: string, message: string, ...args: (string | number | boolean | undefined | null)[]): string;
-}
-
-export interface LoadFunc {
-	(file?: string): LocalizeFunc;
-}
-
-interface NlsBundle {
-	[key: string]: string[];
-}
-
-type SingleFileJsonFormat = string[] | { messages: string[]; keys: string[]; };
-
-export type KeyInfo = string | LocalizeInfo;
-
-interface MetaDataEntry {
-	messages: string[];
-	keys: KeyInfo[];
-}
-
-interface MetadataHeader {
-	id: string;
-	type: string;
-	hash: string;
-	outDir: string;
-}
-
-interface MetaDataFile {
-	[key: string]: MetaDataEntry;
-}
-
-interface TranslationConfig {
-	[extension: string]: string;
-}
-
-interface I18nBundle {
-	version: string,
-	contents: {
-		[module: string]: {
-			[messageKey: string]: string;
-		};
-	}
-}
-
-interface LanguageBundle {
-	header: MetadataHeader;
-	nlsBundle: NlsBundle;
+	return JSON.parse(fs.readFileSync(filename, 'utf8')) as T;
 }
 
 interface VSCodeNlsConfig {
@@ -136,13 +69,11 @@ interface InternalOptions {
 	cacheRoot?: string;
 }
 
-let resolvedLanguage: string;
 let resolvedBundles: {
 	[Key: string]: LanguageBundle | null;
 };
 
 let options: InternalOptions;
-let isPseudo: boolean;
 
 function initializeSettings() {
 	options = { locale: undefined, language: undefined, languagePackSupport: false, cacheLanguageResolution: true, messageFormat: MessageFormat.bundle };
@@ -150,7 +81,6 @@ function initializeSettings() {
 		try {
 			let vscodeOptions = JSON.parse(process.env.VSCODE_NLS_CONFIG) as VSCodeNlsConfig;
 			let language: string | undefined;
-			let locale: string | undefined;
 			if (vscodeOptions.availableLanguages) {
 				let value = vscodeOptions.availableLanguages['*'];
 				if (isString(value)) {
@@ -197,8 +127,7 @@ function initializeSettings() {
 			// Do nothing.
 		}
 	}
-	isPseudo = options.locale === 'pseudo';
-	resolvedLanguage = undefined;
+	setPseudo(options.locale === 'pseudo');
 	resolvedBundles = Object.create(null);
 }
 initializeSettings();
@@ -207,33 +136,6 @@ function supportsLanguagePack(): boolean {
 	return options.languagePackSupport === true && options.cacheRoot !== undefined && options.languagePackId !== undefined && options.translationsConfigFile !== undefined
 		&& options.translationsConfig !== undefined;
 }
-
-function format(message: string, args: any[]): string {
-	let result: string;
-
-	if (isPseudo) {
-		// FF3B and FF3D is the Unicode zenkaku representation for [ and ]
-		message = '\uFF3B' + message.replace(/[aouei]/g, '$&$&') + '\uFF3D';
-	}
-
-	if (args.length === 0) {
-		result = message;
-	} else {
-		result = message.replace(/\{(\d+)\}/g, (match, rest) => {
-			let index = rest[0];
-			let arg = args[index];
-			let replacement = match;
-			if (typeof arg === 'string') {
-				replacement = arg;
-			} else if (typeof arg === 'number' || typeof arg === 'boolean' || arg === void 0 || arg === null) {
-				replacement = String(arg);
-			}
-			return replacement;
-		});
-	}
-	return result;
-}
-
 
 function createScopedLocalizeFunction(messages: string[]): LocalizeFunc {
 	return function (key: any, message: string, ...args: any[]): string {
@@ -254,9 +156,6 @@ function createScopedLocalizeFunction(messages: string[]): LocalizeFunc {
 	};
 }
 
-function localize(key: string | LocalizeInfo, message: string, ...args: any[]): string {
-	return format(message, args);
-}
 
 function resolveLanguage(file: string): string {
 	let resolvedLanguage: string;
@@ -392,7 +291,7 @@ function loadNlsBundleOrCreateFromI18n(header: MetadataHeader, bundlePath: strin
 	let useMemoryOnly: boolean = false;
 	let writeBundle: boolean = false;
 	try {
-		result = JSON.parse(readFileSync(bundle, { encoding: 'utf8', flag: 'r' }));
+		result = JSON.parse(fs.readFileSync(bundle, { encoding: 'utf8', flag: 'r' }));
 		touch(bundle);
 		return result;
 	} catch (err) {
@@ -491,8 +390,9 @@ function tryFindMetaDataHeaderFile(file: string): string {
 }
 
 export function loadMessageBundle(file?: string): LocalizeFunc {
-	if (!file || !readFileSync) {
-		// No file, or fs is not available (bundler empty mock). We are in dev mode.
+	if (!file) {
+		// No file. We are in dev mode. Return the default
+		// localize function.
 		return localize;
 	}
 	// Remove extension since we load json files.
@@ -508,7 +408,7 @@ export function loadMessageBundle(file?: string): LocalizeFunc {
 			let bundle: LanguageBundle = resolvedBundles[bundlePath];
 			if (bundle === undefined) {
 				try {
-					let header: MetadataHeader = JSON.parse(readFileSync(headerFile, 'utf8'));
+					let header: MetadataHeader = JSON.parse(fs.readFileSync(headerFile, 'utf8'));
 					try {
 						let nlsBundle = loadNlsBundle(header, bundlePath);
 						bundle = cacheBundle(bundlePath, nlsBundle ? { header, nlsBundle } : null);
@@ -567,7 +467,6 @@ export function config(opts?: Options): LoadFunc {
 		if (isString(opts.locale)) {
 			options.locale = opts.locale.toLowerCase();
 			options.language = options.locale;
-			resolvedLanguage = undefined;
 			resolvedBundles = Object.create(null);
 		}
 		if (opts.messageFormat !== undefined) {
@@ -577,6 +476,6 @@ export function config(opts?: Options): LoadFunc {
 			options.languagePackSupport = false;
 		}
 	}
-	isPseudo = options.locale === 'pseudo';
+	setPseudo(options.locale === 'pseudo');
 	return loadMessageBundle;
 }
